requirements.txt
streamlit==1.29.0
openai>=2.0.0
rapidfuzz==3.5.2
python-dotenv==1.0.0

